update entity kill + observable draft

+++ amoginarium/logic/entities/_base/_base_entities/_base_logic_entity.py
@@ -49,6 +49,7 @@ class BaseLogicEntity(BaseLogicEntityLike):
     - Groups
     - Update
     - Visibility
+    - Killing
     """
 
     __slots__ = ("_parent", "_children", "_lifetime", "_runtime_buffer", "__id",
@@ -179,14 +180,34 @@ class BaseLogicEntity(BaseLogicEntityLike):
                 group.remove(self)
                 self.__groups.remove(group)
 
-    def _kill(self, killed_by: BaseLogicEntityLike | EllipsisType = ...) -> None:
+    # noinspection PyUnusedLocal, PyMethodMayBeStatic
+    def _before_kill(
+            self,
+            killed_by: BaseLogicEntityLike | EllipsisType = ...,
+            kill_children: bool = True
+    ) -> bool:
+        """
+        Whether the entity can be killed. Called before _kill
+        :param killed_by: who killed this entity
+        :param kill_children: whether to kill children as well recursively
+        :return: Whether the entity kill is accepted.
+        """
+        return True
+
+    def _kill(
+            self,
+            killed_by: BaseLogicEntityLike | EllipsisType = ...,
+            kill_children: bool = True
+    ) -> None:
         """
         Kill entity and all its children
         :param killed_by: who killed this entity
+        :param kill_children: whether to kill children as well recursively
         """
         # kill children first
-        for child in self._children:
-            child.kill()
+        if kill_children:
+            for child in self._children:
+                child.kill()
 
         for group in self.__groups:
             group.remove(self)
@@ -196,15 +217,55 @@ class BaseLogicEntity(BaseLogicEntityLike):
 
         self.__groups.clear()
 
+    def _after_kill(
+            self,
+            killed_by: BaseLogicEntityLike | EllipsisType = ...,
+            kill_children: bool = True,
+            killed: bool = True
+    ) -> None:
+        """
+        Called at the end of kill no matter if the kill was accepted or not
+        :param killed_by: who killed this entity
+        :param kill_children: whether to kill children as well recursively
+        :param killed: Whether the entity kill was accepted or not
+        """
+        ...
+
     @tp.final
-    def kill(self, killed_by: BaseLogicEntityLike | EllipsisType = ...) -> None:
+    def kill(
+            self,
+            killed_by: BaseLogicEntityLike | EllipsisType = ...,
+            kill_children: bool = True,
+            force_kill: bool = False
+    ) -> bool | None:
         """
         Kill entity and all its children
         :param killed_by: who killed this entity
+        :param kill_children: whether to kill children as well as recursively
+        :param force_kill: whether to kill even if before kill returns False
+        :return: Whether the entity was killed or not. May be denied by _before_kill.
+            None if the entity is already dead.
         """
         if self._alive:
-            self._alive = False
-            self._kill(killed_by)
+            killed: bool = False
+
+            kill_entity: bool | None = self._before_kill(
+                killed_by=killed_by,
+                kill_children=kill_children
+            )
+
+            if force_kill or kill_entity is True or kill_entity is None:
+                self._alive = False
+                self._kill(killed_by=killed_by, kill_children=kill_children)
+                killed = True
+
+            self._after_kill(
+                killed_by=killed_by,
+                kill_children=kill_children,
+                killed=killed
+            )
+            return killed
+        return None
 
     # endregion
